bdd: move NominatimEnvironment into separate file
Also cleans up and modernizes the code and adds documentation.
